# docs/openapi-specs.json
{
  "openapi": "3.0.8",
  "info": {
    "title": "🚚 Goldship Logistic API (Golang + Gin)",
    "version": "1.0.0",
    "description": "📃 API documentation for the Goldship Logistic service. Built with :\n\n - [Golang](https://go.dev/)\n\n - [Gin](https://gin-gonic.com/)\n\n - [Google Maps API](https://developers.google.com/maps)\n\n - [Xendit Payment Gateway](https://www.xendit.co/)\n\n - [PostgreSQL](https://www.postgresql.org/)\n\n --- \n\n 🧙‍♂️✨ Wizardly created by [masadamsahid](https://github.com/masadamsahid)"
  },
  "servers": [
    {
      "url": "{host}/",
      "variables": {
        "host": {
          "enum": [
            "http://localhost:8080"
          ],
          "default": "http://localhost:8080"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/branches": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a new branch",
        "description": "Create a new branch. Requires ADMIN or SUPERADMIN role.",
        "tags": [
          "Branches"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Branch created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Branch"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all branches",
        "description": "Retrieve a list of all branches.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of items per page for pagination",
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "tags": [
          "Branches"
        ],
        "responses": {
          "200": {
            "description": "A list of branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Branches retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Branch"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shipments": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a new shipment",
        "description": "Creates a new shipment and returns the details along with a payment invoice URL.",
        "tags": [
          "Shipments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sender_name": {
                    "type": "string",
                    "example": "Raden Raiden"
                  },
                  "sender_phone": {
                    "type": "string",
                    "example": "+6281234567890"
                  },
                  "sender_address": {
                    "type": "string",
                    "example": "Jl. Kaliurang KM 4, Depok, Sleman, DI Yogyakarta, Indonesia. 52811"
                  },
                  "recipient_name": {
                    "type": "string",
                    "example": "Kuda Laper"
                  },
                  "recipient_address": {
                    "type": "string",
                    "example": "Jl. Mangga Dua No. 2, Jakarta, Indonesia. 11110"
                  },
                  "recipient_phone": {
                    "type": "string",
                    "example": "+6289876543210"
                  },
                  "item_name": {
                    "type": "string",
                    "example": "Laptop"
                  },
                  "item_weight": {
                    "type": "number",
                    "format": "float",
                    "example": 2.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shipment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Shipment created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Shipment"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get all shipments",
        "description": "Retrieve a list of all shipments. Requires ADMIN, SUPERADMIN, or COURIER role.",
        "tags": [
          "Shipments"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of items per page for pagination",
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of shipments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Shipments retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Shipment"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProps"
          }
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "user1@mail.com"
          },
          "username": {
            "type": "string",
            "example": "user1"
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "example": "password123"
          },
          "role": {
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRoleEnum"
              }
            ],
            "readOnly": true
          }
        }
      },
      "Branch": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProps"
          }
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Branch 1"
          },
          "phone": {
            "type": "string",
            "example": "+62-81234567890"
          },
          "address": {
            "type": "string",
            "example": "Jl. Jenderal Sudirman No. 5, Jakarta"
          }
        }
      },
      "Shipment": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProps"
          }
        ],
        "properties": {
          "tracking_number": {
            "type": "string",
            "example": "GS-20250829171357SFE5DEIY",
            "readOnly": true
          },
          "sender_id": {
            "type": "integer",
            "example": 1,
            "readOnly": true
          },
          "sender_name": {
            "type": "string",
            "example": "Raden Raiden"
          },
          "sender_phone": {
            "type": "string",
            "example": "+6281234567890"
          },
          "sender_address": {
            "type": "string",
            "example": "Jl. Kaliurang KM 4, Depok, Sleman, DI Yogyakarta, Indonesia. 52811"
          },
          "recipient_name": {
            "type": "string",
            "example": "Kuda Laper"
          },
          "recipient_address": {
            "type": "string",
            "example": "Jl. Mangga Dua No. 2, Jakarta, Indonesia. 11110"
          },
          "recipient_phone": {
            "type": "string",
            "example": "+6289876543210"
          },
          "item_name": {
            "type": "string",
            "example": "Laptop"
          },
          "item_weight": {
            "type": "number",
            "format": "float",
            "example": 2.5
          },
          "distance": {
            "type": "number",
            "format": "float",
            "example": 15.3,
            "readOnly": true
          },
          "base_price": {
            "type": "integer",
            "example": 10000,
            "readOnly": true
          },
          "distance_price": {
            "type": "integer",
            "example": 15000,
            "readOnly": true
          },
          "weight_price": {
            "type": "integer",
            "example": 5000,
            "readOnly": true
          },
          "total_price": {
            "type": "integer",
            "example": 30000,
            "readOnly": true
          },
          "status": {
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/ShipmentStatusEnum"
              }
            ],
            "readOnly": true
          },
          "payment": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Payment"
              }
            ],
            "readOnly": true
          },
          "histories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShipmentHistory"
            },
            "readOnly": true
          }
        }
      },
      "ShipmentHistory": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProps"
          }
        ],
        "properties": {
          "shipment_id": {
            "type": "integer",
            "example": 1,
            "readOnly": true
          },
          "status": {
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/ShipmentStatusEnum"
              }
            ],
            "readOnly": true
          },
          "desc": {
            "type": "string",
            "example": "Shipment created",
            "readOnly": true
          },
          "courier_id": {
            "type": "integer",
            "nullable": true,
            "example": null,
            "readOnly": true
          },
          "branch_id": {
            "type": "integer",
            "nullable": true,
            "example": 1,
            "readOnly": true
          },
          "shipment": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Shipment"
              }
            ],
            "readOnly": true
          },
          "branch": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Branch"
              }
            ],
            "nullable": true,
            "readOnly": true
          },
          "courier": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ],
            "nullable": true,
            "readOnly": true
          }
        }
      },
      "BaseProps": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true,
            "example": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "example": "2025-08-29T17:13:58.621507Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true,
            "example": "2025-08-29T17:13:58.621507Z"
          }
        }
      },
      "Payment": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProps"
          }
        ],
        "properties": {
          "shipment_id": {
            "type": "integer",
            "example": 1,
            "readOnly": true
          },
          "amount": {
            "type": "integer",
            "example": 30000,
            "readOnly": true
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true,
            "example": "2025-08-29T17:13:58.621507Z"
          },
          "expired_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true,
            "example": "2025-08-29T17:13:58.621507Z"
          },
          "invoice_id": {
            "type": "string",
            "example": "invoice-123",
            "readOnly": true
          },
          "external_id": {
            "type": "string",
            "example": "external-123",
            "readOnly": true
          },
          "invoice_url": {
            "type": "string",
            "example": "https://invoice.url/123",
            "readOnly": true
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatusEnum"
              }
            ],
            "readOnly": true
          }
        }
      },
      "UserRoleEnum": {
        "type": "string",
        "enum": [
          "SUPERADMIN",
          "ADMIN",
          "COURIER",
          "CUSTOMER"
        ],
        "example": "CUSTOMER"
      },
      "ShipmentStatusEnum": {
        "type": "string",
        "enum": [
          "PENDING_PAYMENT",
          "READY_TO_PICKUP",
          "PICKED_UP",
          "IN_TRANSIT",
          "DELIVERED",
          "CANCELLED"
        ],
        "example": "PENDING_PAYMENT"
      },
      "PaymentStatusEnum": {
        "type": "string",
        "enum": [
          "PENDING",
          "PAID",
          "EXPIRED",
          "CANCELLED"
        ],
        "example": "PENDING"
      },
...
    }
  }
}